+++ project/src/ListBroking/AppBundle/Service/Helper/StatisticsServiceInterface.php
@@ -5,6 +5,8 @@
  */
 namespace ListBroking\AppBundle\Service\Helper;
 
+use ListBroking\AppBundle\Model\AudiencesFilter;
+
 interface StatisticsServiceInterface
 {
     /**
@@ -15,4 +17,24 @@ interface StatisticsServiceInterface
      * @return bool
      */
     public function generateStatisticsQuery ($data);
+
+    /**
+     * Calculates audiences based on:
+     * - Owner
+     * - Country
+     * - District
+     * - Category
+     * - Gender
+     * - Age
+     */
+    public function calculateAudiences(): void;
+
+    /**
+     * Get audiences based on the given filter
+     *
+     * @param AudiencesFilter $filter
+     *
+     * @return array
+     */
+    public function getAudiences(AudiencesFilter $filter): array;
 }